# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "OUTLIER API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/listings/{listing_id}": {
   "get": {
    "tags": [
     "listings"
    ],
    "summary": "Get Listing",
    "operationId": "get_listing_api_listings__listing_id__get",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "listings"
    ],
    "summary": "Patch Listing",
    "operationId": "patch_listing_api_listings__listing_id__patch",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ListingPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "listings"
    ],
    "summary": "Delete Listing",
    "operationId": "delete_listing_api_listings__listing_id__delete",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/listings/{listing_id}/comps": {
   "get": {
    "tags": [
     "comparables"
    ],
    "summary": "List Comps",
    "operationId": "list_comps_api_listings__listing_id__comps_get",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "comparables"
    ],
    "summary": "Add Comp",
    "operationId": "add_comp_api_listings__listing_id__comps_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CompIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/comps/import": {
   "post": {
    "tags": [
     "comparables"
    ],
    "summary": "Import Comps",
    "operationId": "import_comps_api_listings__listing_id__comps_import_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_comps_api_listings__listing_id__comps_import_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/comps/{comp_id}": {
   "patch": {
    "tags": [
     "comparables"
    ],
    "summary": "Patch Comp",
    "operationId": "patch_comp_api_listings__listing_id__comps__comp_id__patch",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     },
     {
      "name": "comp_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Comp Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CompPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "comparables"
    ],
    "summary": "Delete Comp",
    "operationId": "delete_comp_api_listings__listing_id__comps__comp_id__delete",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     },
     {
      "name": "comp_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Comp Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/comps/search": {
   "post": {
    "tags": [
     "comparables"
    ],
    "summary": "Search Comps",
    "description": "Query a live sold-data provider (requires credentials + eBay approval). Results are NOT saved unless save=true.",
    "operationId": "search_comps_api_listings__listing_id__comps_search_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CompSearchIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/valuation/override": {
   "post": {
    "tags": [
     "comparables"
    ],
    "summary": "Override Valuation",
    "operationId": "override_valuation_api_listings__listing_id__valuation_override_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ValuationOverride"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/valuation/recalculate": {
   "post": {
    "tags": [
     "comparables"
    ],
    "summary": "Recalc Valuation",
    "operationId": "recalc_valuation_api_listings__listing_id__valuation_recalculate_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/finance": {
   "post": {
    "tags": [
     "comparables"
    ],
    "summary": "Finance",
    "operationId": "finance_api_listings__listing_id__finance_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FinanceQuery"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_import_comps_api_listings__listing_id__comps_import_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_import_comps_api_listings__listing_id__comps_import_post"
   },
...
   "CompIn": {
    "properties": {
     "title": {
      "type": "string",
      "maxLength": 500,
      "minLength": 1,
      "title": "Title"
     },
     "marketplace": {
      "type": "string",
      "title": "Marketplace",
      "default": "ebay"
     },
     "url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Url"
     },
     "sold_date": {
      "title": "Sold Date"
     },
     "price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price"
     },
     "shipping_included": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shipping Included"
     },
     "shipping_amount": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shipping Amount"
     },
     "is_sold": {
      "type": "boolean",
      "title": "Is Sold",
      "default": true
     },
     "accepted_offer": {
      "type": "boolean",
      "title": "Accepted Offer",
      "default": false
     },
     "comp_type": {
      "type": "string",
      "title": "Comp Type",
      "default": "same_maker"
     },
     "similarity": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Similarity",
      "default": 0.6
     },
     "evidence_quality": {
      "type": "string",
      "title": "Evidence Quality",
      "default": "medium"
     },
     "condition": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Condition"
     },
     "differences": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Differences"
     },
     "source": {
      "type": "string",
      "title": "Source",
      "default": "manual"
     }
    },
    "type": "object",
    "required": [
     "title"
    ],
    "title": "CompIn"
   },
   "CompPatch": {
    "properties": {
     "comp_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Comp Type"
     },
     "similarity": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 1.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Similarity"
     },
     "evidence_quality": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Evidence Quality"
     },
     "price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price"
     },
     "is_sold": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Sold"
     },
     "differences": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Differences"
     },
     "condition": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Condition"
     }
    },
    "type": "object",
    "title": "CompPatch"
   },
   "CompSearchIn": {
    "properties": {
     "query": {
      "type": "string",
      "title": "Query"
     },
     "provider": {
      "type": "string",
      "title": "Provider",
      "default": "ebay_marketplace_insights"
     },
     "limit": {
      "type": "integer",
      "title": "Limit",
      "default": 20
     },
     "save": {
      "type": "boolean",
      "title": "Save",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "query"
    ],
    "title": "CompSearchIn"
   },
...
   "FinanceQuery": {
    "properties": {
     "bid": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bid"
     },
     "platform": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Platform"
     },
     "resale_price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Resale Price"
     },
     "overrides": {
      "additionalProperties": true,
      "type": "object",
      "title": "Overrides"
     },
     "sensitivity": {
      "type": "boolean",
      "title": "Sensitivity",
      "default": true
     }
    },
    "type": "object",
    "title": "FinanceQuery"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ListingPatch": {
    "properties": {
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "seller": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seller"
     },
     "domain": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Domain"
     },
     "condition_text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Condition Text"
     },
     "measurements": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Measurements"
     },
     "other_costs": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "number"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Other Costs"
     },
     "assumptions": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Assumptions"
     },
     "user_notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Notes"
     },
     "archived": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Archived"
     },
     "status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     },
     "source_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Url"
     }
    },
    "type": "object",
    "title": "ListingPatch"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "ValuationOverride": {
    "properties": {
     "expected": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Expected"
     },
     "conservative": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Conservative"
     },
     "optimistic": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Optimistic"
     },
     "note": {
      "type": "string",
      "title": "Note",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "expected"
    ],
    "title": "ValuationOverride"
   },
...
  }
 }
}